pltChromeLogo/requirement.txt
PyYAML==5.3
matplotlib==3.1.3